refactor: eliminate `babylon-bitcoin` crate (#197)
This PR removes `packages/bitcoin`, as the crate is no longer needed.
The code it contained was consumer-specific and not reusable across
multiple consumers.

Now we only define a custom `BitcoinNetwork` type in `btc-light-client`
to mirror `bitcoin::Network` where necessary, specifically to implement
the missing `JsonSchema` trait and to persist network state at
boundaries of storage and execution. In all other cases, we default to
using `bitcoin::Network` directly for better leveraging `rust-bitcoin`.