chore: restructure btcstaking pkg to align w/ Go (#204)
As we are trying to replicate the Golang implementation of the BTC
staking repository, it will be much more easier to compare those if they
follow a similar directory structure.

This PR does the following:
* Aligns the file of the package with the Golang btcstaking library
* Moves the schnorr adaptor signature functionality to its own package

Note: the change only moves code around, it does not actually modify
anything. If there are modifications suggested, we can implement them in
a subsequent PR.