fix(btc-light-client): improve integration test and CI (#211)
I just noticed the integration test in btc-light-client is broken, this
PR fixes it by moving it to the right location. The updated integration
test now matches the other ones in the other contracts. ~~It also
updates the `local-tests.yml` workflow to run all workspace tests,
ensuring better test coverage~~.