fix(btc-light-client): store the base header data (#223)
Gai discovered this bug while working on updating the contracts e2e
tests in babylon-sdk, see #222 for more context.

Close #222

One thing I noticed: Currently, we store the entire header for the BTC
tip, which might be unnecessary as we already have the tip header stored
in the `BTC_HEADERS` storage. We may only store the metadata
(header_hash, header_height) for the BTC tip.