finality: handle public randomness misalignment (#265)
This PR aims to resolve the misalignment listed in #262.

One important thing in this PR is the introduction of the
`PublicRandomnessCommitMsg` type which mirrors the original Go
implementation. I strongly recommend we do the same to the others, i.e.,
we only unpack the fields in the tx interface layer, once the msg
crosses the boundary, we should immediately convert them into the types
like `PublicRandomnessCommitMsg`, making it easier to reason about and
align with the original Go impl and audit.

TODO:

- [ ] Port the Go tests.

Close #262