bugfix: out of gas on big batch (#349)
## Description

This PR fixes the default size for
`MAX_FINALIZED_REWARDED_BLOCKS_PER_END_BLOCK`. The previous default of
`1_000` was too large for the default block gas limit and caused runtime
errors when sending the `EndBlock` message to the BTC finality contract.
With `1_000`, we observed panics such as:
```
10:16PM ERR Failed to send EndBlock message to BTC finality contract error="contract call to bbnc1ghd753shjuwexxywmgs4xz7x2q732vcnkm6h2pyv9s6ah3hylvrqv7u2f2 panicked: {Wasm engine function execution}, gas_used: 5000000" module=x/babylon
```
To prevent this, we picked a more conservative default range that fits
safely within the block gas limit, ensuring EndBlock execution will not
exceed gas and panic under normal conditions.

---------

Co-authored-by: Runchao Han <me@runchao.rocks>